mkdir unifiedload and move LoadType to it
