+++ fe/fe-core/src/main/java/org/apache/doris/analysis/MysqlLoadStmt.java
@@ -26,6 +26,7 @@ import org.apache.doris.common.AnalysisException;
 import org.apache.doris.common.Config;
 import org.apache.doris.common.DdlException;
 import org.apache.doris.common.UserException;
+import org.apache.doris.load.unifiedload.LoadType;
 
 import java.io.File;
 import java.io.IOException;
